% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_playlists.r
\name{plot_complexity}
\alias{plot_complexity}
\title{Plot Complexity Distribution}
\usage{
plot_complexity(analysis_result)
}
\arguments{
\item{analysis_result}{Result from analyze_playlist()}
}
\description{
Plot Complexity Distribution
}
