% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_playlists.r
\name{plot_sentiment}
\alias{plot_sentiment}
\title{Plot Sentiment Distribution}
\usage{
plot_sentiment(analysis_result)
}
\arguments{
\item{analysis_result}{Result from analyze_playlist()}
}
\description{
Plot Sentiment Distribution
}
