% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_playlists.r
\name{plot_top_complex}
\alias{plot_top_complex}
\title{Plot Top 10 Most Complex Songs}
\usage{
plot_top_complex(analysis_result)
}
\arguments{
\item{analysis_result}{Result from analyze_playlist()}
}
\description{
Plot Top 10 Most Complex Songs
}
