% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlyrics.R
\name{preview_lyrics}
\alias{preview_lyrics}
\title{Preview Lyrics}
\usage{
preview_lyrics(x, n = 5)
}
\arguments{
\item{x}{A lyrics_lab object}

\item{n}{Number of lines to show. Default 5}
}
\description{
Shows metadata and the first few lines of a lyrics_lab object.
}
\examples{
\dontrun{
lyrics <- get_lyrics("Taylor Swift", "Anti-Hero")
preview_lyrics(lyrics)
preview_lyrics(lyrics, n = 10)
}
}
