% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_connection.r
\name{spotify_auth}
\alias{spotify_auth}
\title{Connect to Spotify API}
\usage{
spotify_auth(client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{Your Spotify Client ID}

\item{client_secret}{Your Spotify Client Secret}
}
\value{
Access token (saved to environment)
}
\description{
Authenticates with Spotify and gets user authorization.
You'll need to create a Spotify App first at https://developer.spotify.com/dashboard
}
\examples{
\dontrun{
spotify_auth(
  client_id = "your_client_id",
  client_secret = "your_client_secret"
)
}

}
