% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlyrics.R
\name{view_lyrics}
\alias{view_lyrics}
\title{View Full Lyrics}
\usage{
view_lyrics(x)
}
\arguments{
\item{x}{A lyrics_lab object}
}
\description{
Prints all lines of a lyrics_lab object.
}
\examples{
\dontrun{
lyrics <- get_lyrics("Taylor Swift", "Anti-Hero")
view_lyrics(lyrics)
}
}
